+++ backend/index.js
@@ -1,0 +1,12 @@
+const express = require('express');
+const app = require('./src/app');
+
+PORT = process.env.PORT | 3000;
+
+app.listen(PORT, () => {
+    console.log(`Server is running on http://localhost:${PORT}`);
+});
+
+
+
+
